Removed WinForms

+++ NodeEditorCore/AnimateForms/Easings.cs
@@ -13,146 +13,146 @@ namespace AnimateForms.Core
 
         // Easing equations pulled from http://gizma.com/easing/
 
-        public static float Linear(float t, float b, float c, float d)
+        public static double Linear(double t, double b, double c, double d)
         {
             return c * t / d + b;
         }
 
-        public static float QuadIn(float t, float b, float c, float d)
+        public static double QuadIn(double t, double b, double c, double d)
         {
             t /= d;
             return c * t * t + b;
         }
 
-        public static float QuadOut(float t, float b, float c, float d)
+        public static double QuadOut(double t, double b, double c, double d)
         {
             t /= d;
             return -c * t * (t - 2) + b;
         }
 
-        public static float QuadInOut(float t, float b, float c, float d)
+        public static double QuadInOut(double t, double b, double c, double d)
         {
             t /= d / 2;
-            if (t < 1) return (float)(c / 2 * t * t + b);
+            if (t < 1) return (double)(c / 2 * t * t + b);
             t--;
             return -c / 2 * (t * (t - 2) - 1) + b;
         }
 
-        public static float CubicIn(float t, float b, float c, float d)
+        public static double CubicIn(double t, double b, double c, double d)
         {
             t /= d;
             return c * t * t * t + b;
         }
 
-        public static float CubicOut(float t, float b, float c, float d)
+        public static double CubicOut(double t, double b, double c, double d)
         {
             t /= d;
             t--;
             return c * (t * t * t + 1) + b;
         }
 
-        public static float CubicInOut(float t, float b, float c, float d)
+        public static double CubicInOut(double t, double b, double c, double d)
         {
             t /= d / 2;
-            if (t < 1) return (float)(c / 2 * t * t * t + b);
+            if (t < 1) return (double)(c / 2 * t * t * t + b);
             t -= 2;
             return c / 2 * (t * t * t + 2) + b;
         }
 
-        public static float QuartIn(float t, float b, float c, float d)
+        public static double QuartIn(double t, double b, double c, double d)
         {
             t /= d;
             return c * t * t * t * t + b;
         }
 
-        public static float QuartOut(float t, float b, float c, float d)
+        public static double QuartOut(double t, double b, double c, double d)
         {
             t /= d;
             t--;
             return -c * (t * t * t * t - 1) + b;
         }
 
-        public static float QuartInOut(float t, float b, float c, float d)
+        public static double QuartInOut(double t, double b, double c, double d)
         {
             t /= d / 2;
-            if (t < 1) return (float)(c / 2 * t * t * t * t + b);
+            if (t < 1) return (double)(c / 2 * t * t * t * t + b);
             t -= 2;
             return -c / 2 * (t * t * t * t - 2) + b;
         }
 
-        public static float QuintIn(float t, float b, float c, float d)
+        public static double QuintIn(double t, double b, double c, double d)
         {
             t /= d;
             return c * t * t * t * t * t + b;
         }
 
-        public static float QuintOut(float t, float b, float c, float d)
+        public static double QuintOut(double t, double b, double c, double d)
         {
             t /= d;
             t--;
             return c * (t * t * t * t * t + 1) + b;
         }
 
-        public static float QuintInOut(float t, float b, float c, float d)
+        public static double QuintInOut(double t, double b, double c, double d)
         {
             t /= d / 2;
-            if (t < 1) return (float)(c / 2 * t * t * t * t * t + b);
+            if (t < 1) return (double)(c / 2 * t * t * t * t * t + b);
             t -= 2;
             return c / 2 * (t * t * t * t * t + 2) + b;
         }
 
-        public static float SinIn(float t, float b, float c, float d)
+        public static double SinIn(double t, double b, double c, double d)
         {
-            return (float)(-c * Math.Cos(t / d * (Math.PI / 2)) + c + b);
+            return (double)(-c * Math.Cos(t / d * (Math.PI / 2)) + c + b);
         }
 
-        public static float SinOut(float t, float b, float c, float d)
+        public static double SinOut(double t, double b, double c, double d)
         {
-            return (float)(c * Math.Sin(t / d * (Math.PI / 2)) + b);
+            return (double)(c * Math.Sin(t / d * (Math.PI / 2)) + b);
         }
 
-        public static float SinInOut(float t, float b, float c, float d)
+        public static double SinInOut(double t, double b, double c, double d)
         {
-            return (float)(-c / 2 * (Math.Cos(Math.PI * t / d) - 1) + b);
+            return (double)(-c / 2 * (Math.Cos(Math.PI * t / d) - 1) + b);
         }
 
-        public static float ExpIn(float t, float b, float c, float d)
+        public static double ExpIn(double t, double b, double c, double d)
         {
-            return (float)(c * Math.Pow(2, 10 * (t / d - 1)) + b);
+            return (double)(c * Math.Pow(2, 10 * (t / d - 1)) + b);
         }
 
-        public static float ExpOut(float t, float b, float c, float d)
+        public static double ExpOut(double t, double b, double c, double d)
         {
-            return (float)(c * (-Math.Pow(2, -10 * t / d) + 1) + b);
+            return (double)(c * (-Math.Pow(2, -10 * t / d) + 1) + b);
         }
 
-        public static float ExpInOut(float t, float b, float c, float d)
+        public static double ExpInOut(double t, double b, double c, double d)
         {
             t /= d / 2;
-            if (t < 1) return (float)(c / 2 * Math.Pow(2, 10 * (t - 1)) + b);
+            if (t < 1) return (double)(c / 2 * Math.Pow(2, 10 * (t - 1)) + b);
             t--;
-            return (float)(c / 2 * (-Math.Pow(2, -10 * t) + 2) + b);
+            return (double)(c / 2 * (-Math.Pow(2, -10 * t) + 2) + b);
         }
 
-        public static float CircIn(float t, float b, float c, float d)
+        public static double CircIn(double t, double b, double c, double d)
         {
             t /= d;
-            return (float)(-c * (Math.Sqrt(1 - t * t) - 1) + b);
+            return (double)(-c * (Math.Sqrt(1 - t * t) - 1) + b);
         }
 
-        public static float CircOut(float t, float b, float c, float d)
+        public static double CircOut(double t, double b, double c, double d)
         {
             t /= d;
             t--;
-            return (float)(c * Math.Sqrt(1 - t * t) + b);
+            return (double)(c * Math.Sqrt(1 - t * t) + b);
         }
 
-        public static float CircInOut(float t, float b, float c, float d)
+        public static double CircInOut(double t, double b, double c, double d)
         {
             t /= d / 2;
-            if (t < 1) return (float)(-c / 2 * (Math.Sqrt(1 - t * t) - 1) + b);
+            if (t < 1) return (double)(-c / 2 * (Math.Sqrt(1 - t * t) - 1) + b);
             t -= 2;
-            return (float)(c / 2 * (Math.Sqrt(1 - t * t) + 1) + b);
+            return (double)(c / 2 * (Math.Sqrt(1 - t * t) + 1) + b);
         }
     }
 }
